Changed so that several amounts of KE can be lost during inel scattering
I added to the Scatterer class, the attribute loss_fn. It is a list of lists. the sub-lists consist of [oorbability, energy_loss]
I added the method to the Scatterer class called getDeltaKE.
It randomly selects an item from the loss_fn and compares to a random number.
Then returns the amount of KE lost.